requirements.txt
django==2.2.13
eventlet==0.30.0
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2018.7
six==1.10.0
docker==4.4.1